workers/terminal/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mss==9.0.1
pyautogui==0.9.54
pynput==1.7.6
pillow==10.2.0
python-dotenv==1.0.0
